Make ComputerPlayer's randomness configurable
The option for predictable ComputerPlayer selections makes it possible to
test the ComputerPlayer's behavior while preserving randomness in
the game environment. The ComputerPlayer selects
spaces based off of a RandomNumberGenerator. The RandomNumberGenerator takes
in a rand method that either returns
predictable output based on a SEED environment variable that is
provided at runtime, or random output based on a seed randomly selected
by Ruby.